+++ tests/SideBySide/QueryTests.cs
@@ -124,7 +124,7 @@ create table query_invalid_sql(id integer not null primary key auto_increment);"
 			}
 		}
 
-		[UnbufferedResultSetsFact]
+		[SkippableFact(ConfigSettings.UnbufferedResultSets)]
 		public async Task MultipleReaders()
 		{
 			using (var cmd = m_database.Connection.CreateCommand())
@@ -168,11 +168,7 @@ create table query_invalid_sql(id integer not null primary key auto_increment);"
 			}
 		}
 
-#if BASELINE
-		[Fact(Skip = "Does not support BufferResultSets")]
-#else
-		[Fact]
-#endif
+		[SkippableFact(Baseline = "Does not support BufferResultSets")]
 		public async Task MultipleBufferedReaders()
 		{
 			var csb = AppConfig.CreateConnectionStringBuilder();
@@ -440,11 +436,7 @@ insert into query_null_parameter (id, value) VALUES (1, 'one'), (2, 'two'), (3, 
 			}
 		}
 
-#if BASELINE
-		[Fact(Skip = "http://bugs.mysql.com/bug.php?id=82292")]
-#else
-		[Fact]
-#endif
+		[SkippableFact(Baseline = "http://bugs.mysql.com/bug.php?id=82292")]
 		public void DapperNullableBoolNullFirst()
 		{
 			// adapted from https://github.com/StackExchange/dapper-dot-net/issues/552
@@ -458,12 +450,8 @@ insert into query_null_parameter (id, value) VALUES (1, 'one'), (2, 'two'), (3, 
 			Assert.True(rows[1].IsBold);
 			Assert.Null(rows[2].IsBold);
 		}
-
-#if BASELINE
-		[Fact(Skip = "https://bugs.mysql.com/bug.php?id=78760")]
-#else
-		[Fact]
-#endif
+		
+		[SkippableFact(Baseline = "https://bugs.mysql.com/bug.php?id=78760")]
 		public void TabsAndNewLines()
 		{
 			m_database.Connection.Execute(@"drop table if exists query_tabs;
@@ -738,11 +726,7 @@ insert into enum_test (id, value) VALUES (1002, 'no'), (1003, 'yes');
 			}
 		}
 
-		[Fact
-#if BASELINE
-			(Skip = "https://bugs.mysql.com/bug.php?id=84701")
-#endif
-		]
+		[SkippableFact(Baseline = "https://bugs.mysql.com/bug.php?id=84701")]
 		public void Int64EnumParameter()
 		{
 			m_database.Connection.Execute(@"drop table if exists long_enum_test;